refactor : 판매입찰 내역에 대한 조회 반환값 수정 :: Sell -> SellPriceToQuantityResponseDto(판매입찰가격수량 내역)
