+++ src/main/java/bc1/gream/domain/sell/dto/response/SellPriceToQuantityResponseDto.java
@@ -1,0 +1,11 @@
+package bc1.gream.domain.sell.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record SellPriceToQuantityResponseDto(
+    Long sellPrice,
+    Long quantity
+) {
+
+}
